Create a place for mining functionality (#64)
Move block creation into a new package, mining. The rest of the block mining code in commands/miner will move there in a followup so that it can be re-used and we're only left with glue in the command implementation. Introduce some test helpers along the way -- we should keep code duplication to a minimum.

This is the first small step towards https://github.com/filecoin-project/go-filecoin/issues/16.