centralize mining logic (#67)
Introduce mining.Worker which can .Mine(). I avoided the name mining.Miner to avoid confusion with a person who is a "miner". I've trimmed what were fat dependencies between the mining code and the rest of the system and I've introduced some separation of duties. Specifically the Worker will handle scheduling and interaction with other bits of the system and the BlockGenerator will do the heavy lifting. We'll refactor as needed.