Implement mining (#82)
Fill in the gaps in Worker and wire it up to the node.

I chose not to move to a builder pattern for Worker, can follow on with that if you think it would be useful. Also, the worker code is close to generic, it would be straightforward to factor it out so that it is a worker for anything that gets inputs and outputs via channels. 

This is work towards #16 and extends #73 which was abandoned.