handle errors better in msg processing (take 2) (#106)
The main thing here is distinguishing different types of errors that can happen in message processing. We now use revert error to signal a vm error that rolls state back but does not indicate a failed message. We have a fault error that signals that the system is badly broken and that we should stop doing whatever it is that we are doing (we just panic now, but the door is open to something better). 

The message processing model here draws on ideas from ethereum. I find the ethereum model for message processing to be pretty obtuse, for example see all the explanation required on ApplyMessage. But, there you have it.

There is also a quick-and-easy implementation of state tree snapshotting and reversion. I think the API is right but obviously the implementation is not ideal. I like what ethereum has done in this particular area. The quick and easy implementation surfaced bugs in the hamt implementation which I'm fixing in ipfs/go-hamt-ipld#2. If you patch those changes in you can uncomment the snapshot/revert and its tests and see that they all pass.

Believe it or not this is a step towards #16. The thing I'm working towards is being able to correctly manage messages in the pool that make it into blocks or fail. To do that we needed to distinguish between different types of failures.

This way of handling errors is still being refined. Continue conversation at https://github.com/filecoin-project/go-filecoin/issues/119