keep message pool up to date, attempt 2 (#173)
When we adopt a new best chain update the message pool accordingly. Remove those messages that were in the old chain rooted at the common ancestor and add back those that are not found in the new chain.

Closes #16. This is all we need for catfood, other issues spun out for dogfood.

This change originally went in as 615ae22 then was reverted in 3f52965 because CI started failing. The 9d328be commit (9d328be) has the changes to fix the failing test (checking ok status on receiving channels -- the changes to message pool are just out of prudence when i thought the problem was there).