clarify ApplyMessage interface (#261)
In working to properly support nonces (#155 ) it became clear that we need to clean up and clarify the contract with ApplyMessage and ProcessBlock. We couldn't easily add the cases we needed without enforcing a separation between vm errors and ApplyMessage errors, making explicit intent, and separating the step of attempting to apply the message from its side effects. These changes are a big step in that direction. There are several additional, parallel PRs coming that mark more incremental progress towards #155 and once they land I can land the step of turning on nonce checking, which is a big PR all by itself.

This is work mostly done while pairing with @clkunzang . It also contributes significantly towards #119