more nonce building blocks: function to order pending messages by nonce (#263)
This is work towards #155 . Here we introduce a function that sorts the pending messages by nonce. We'll eventually need more sophisticated and efficient sorting, but this works for now.

Also move fake actor into a testing.go and export it so we can use it in BlockGenerator tests. We're going to supplant the behavioral tests in block_generator_test.go with more contractual tests and to do that we'll want to mimic processblock's tests which use fake actors.