implement primitive network bootstrapping (#500)
Introduce a filnet.Bootstrapper based on the ipfs bootstrapping pattern. Currently the bootstrapper won't actually do anything (Bootstrapper.MinPeerThreshold == 0) because we don't have bootstrap nodes to connect to. In the process created a new package for network-related code, filnet and moved a few functions there (and added a unit test for them).

Closes https://app.zenhub.com/workspace/o/filecoin-project/go-filecoin/issues/454