Support for multi-block TipSets
With this change the filecoin node now tracks the heaviest TipSet which can
contain more than one block.  Chain validation now works for histories with
multi-block TipSets.  To accomplish this a new state aggregation function
is introduced to calculated the child state of a TipSet.  The base state
used by mining is likewise calculated with this function, relative the
heaviest TipSet that is now tracked by the ChainManager.

The assumption that a chain will have a single block at each epoch and one
bestBlock at any time is replaced throughout the codebase and rewritten to
be compatible with a chain that can have multiple winning blocks in a given
epoch. This is an extensive change originally taking place over 32 separate
commits. This change updates the chainManager, processor functions, and the
message pool within core as well as functions within the mining worker and
blockGenerator and the filecoin node.  Superficial changes propagate to
commands and go-fakecoin.  Changes hit the tests of all listed modules and
update several test helpers.  The TipSet type withing tip_index.go is
augmented with many new functions.