make TokenAmount AttoFIL, a count of attofilecoin (#560)
This is work towards #340 in preparation to switch from big.Int's byte encoding to our leb128 encoding.

This change does not modify what you give to and (mostly) see in commands for token amounts, which is units of filecoin. This change is already enormous so that can be a followon for another day. I don't plan to get to it right now.