First pass at EC chain weighting. This commit adds the notion of miner power to the chain weight calculation. Block fields and chain manager logic are updated accordingly.
Due to tons of existing tests that rely on mining without setting up a
valid state tree this commit introduces a mockMining mode to the node and
CLI.  mockMining mode causes the chain manager to process blocks with
a simplified view of the power table fo tests.  New tests exercise the
rational chain weights.