Refactor Mining Code
Before this mining was split roughly into the Worker and BlockGenerator
interfaces.  This commit changes this to a Scheduler and Worker interface.

Scheduler -- we are now creating a more complex mining strategy that
involves timing.  The Scheduler interface is where these strategies fit
into the mining code.  This interface has a Start() method that begins
mining and provides input and output channels to the rest of the system
just like the old Worker interface.  Schedulers schedule mining work for
a Worker.

Worker -- as before Worker is the thing that actually mines.  However in
this commit Worker is now seperate from the thing that Starts mining
(that is the Scheduler).  Workers do work with the interface Mine method.

BlockGenerator -- the BlockGenerator has been absorbed into the Worker.
Here is my rational.
1. This facilitates making the code easier to follow with respect to
@whyrusleeping's issues.  We are now one interface and one object fewer
in the mining code.
2. The seperation is no longer justified.  In my view now that the worker
is no longer responsible for scheduling, there is enough room in the
abstraction for block generation as a member function.  Looking back at
the code of Mine I was not convinced this interface is worth the trouble.
It seems likely that any alternate implementation of Generate would warrant
an alternate implementation of Mine.  I don't forsee lots of swapping
implementations across the BlockGenerator interface and don't think
testing Mine against an abstract Generate buys us that much
3. The seperation is actively getting in the way.  worker.Mine has been
using fake power values for a while.  The way to fix this is to use the
powerTableView and getStateTree dependencies used by the block generator.
It makes far more sense for one thing to share these dependencies than
to inject them into two things.  With this refactor we are much closer
to getting actual power values.

Lots of tests have changed but this commit has kept most of the old
behavioral coverage.  Moving the seam to Scheduler/Worker is about
as fine-grained as the old seam between Worker/BlockGenerator.
Lots of Scheduler tests are missing because the timingScheduler is still
broken.  Some of the behavior moved into the scheduler from the Mine method,
e.g. null block incrementing is not yet tested.