Pretty Print/Parse AttoFIL in JSON output
We change the AttoFIL#String method to output AttoFIL as a decimal
number of FIL. This method is used to displayed quantities of AttoFIL in
JSON to users of the CLI. This format is loss-less, since the decimal
representation is exact.

We also change NewAttoFILFromFILString to accept a base-10 decimal
representation of FIL as a string. Combined with the above, we get a
completely loss-less (un)marshaling as JSON.