Clean up properly on SIGTERM
We clean up properly when the filecoin daemon is terminated (via killall
or kill). This clean up removes the api and repo.lock files in the repo,
preventing a command from erroneously thinking a daemon is running.

We make the testing a little more robust by actually checking for any
(unexpected) Errors in the logs as well. Warnings are no longer checked,
as these are generated during normal operation of the filecoin node and
are frequently outside of the control of this repository