pull message waiter off of node into its own package (#1224)
This is work towards #1223. Big picture is captured in the issue but basically in order to write sane tests for SendMessageAndWait we want (1) SendMessageAndWait to not depend on the whole node and (2) the wait implementation to not be in node.

In this step the code is moved into its own package. I kept the logic and tests the same as they were with the exception of a bug I noticed when moving the code, which I fixed and added a test for (TestWaitRespectsContextCancel). The new chain/init is just the chain-related subset of code i moved from node/init. Notice how after this change the chain wait tests now only depend on the ChainStore and not on the Node. This enables the follow up steps which otherwise would create a circular dependency.