Prevent Bad Nodes From Joining Cluster
We prevent bad nodes from joining the test cluster by basing the genesis
block seed on the unix time of the build. This means that every new
build of the test cluster will require the genesis block from the latest
build.

We do this by:
- Allowing the `genesisfile` argument in `init` to be an HTTP(S) URL for
  downloading the genesis block.
- Putting a simple HTTP server in the project to serve up the genesis
  file.

We additionally cause the libp2p connection to terminate if one connects
to a peer with a bad genesis block. This is a change to the hello
protocol.