move proof generation to SectorBuilder
To create a proof-of-spacetime, you need access to the replicas. The previously-existing RustProver was a pure function; it didn't know how to read from the disk and it certainly didn't know how to map comm-r's to sealed sectors.

This PR splits the proof-generation stuff out of the RustProver, moving it to the SectorBuilder (we probably need a new name for that interface). RustProver has been renamed RustVerifier, and the Prover interface has been renamed to Verifier.