refactor PoSt generation/verification to be a little closer towards reality  (#1593)
* proof-of-spacetime generation/verification refactor

- don't verify post as part of SubmitPoSt
- add constants for various proof-related things
- go-filecoin generates challenge seed, not challenge (challenge generation happens in rust-proofs)
- verify_post requires commrs, faults, and challenge seed (in addition to proof-bytes)

* rust prover becomes rust verifier

* Prover -> Verifier

* move proof generation to SectorBuilder

To create a proof-of-spacetime, you need access to the replicas. The previously-existing RustProver was a pure function; it didn't know how to read from the disk and it certainly didn't know how to map comm-r's to sealed sectors.

This PR splits the proof-generation stuff out of the RustProver, moving it to the SectorBuilder (we probably need a new name for that interface). RustProver has been renamed RustVerifier, and the Prover interface has been renamed to Verifier.

* update rust-proofs submodule

* prover -> verifier

* verify proof-of-spacetime immediately after generating it

* test as slice

* add ErrInvalidPost to singleton Errors

* fix typo

* introduce new proofsy types