Proofs download: speed ups
This prevents the download of the full ~50MB of RELEASE_URL to /dev/null only
to find the effective_url (ASSET_URL) and re-download the file (I am not sure
why it's done this way but I guess there is a reason).

It also avoids re-downloading a proof file that has already been downloaded
by saving them with a unique file.

100% speed up in  pre-compiled proofs installation
1000000% speed up in pre-compiled proofs re-installation

Also, mention that rust, jq and curl are needed in the readme.