prevents groth parameters from being generated on bootstrap symlink v9-zigzag-proof-of-replication-52431242c129794fe51d373ae29953f2ff52abd94c78756e318ce45f3e4946d8 to param.out.
If there exists a file called `params.out`, it will be treated specially. It is the name of the 'official' groth params file to be used by seal. If it exists, it will be loaded once and kept in memory.
fixes the problem temporarily.

should find a less brittle solution in the near future