Coalesce storage deal stores
We coalesce the storage deal stores for both miners and clients so they
can easily be queried by each. This allows miners to query the vouchers
they have for various storage deals, along with the state of those
deals.

We also push the deal storage mechanism into plumbing, though it may be
moved shortly with the upcoming protocol refactor.

There is also some simple cleanup such as:
- Removing some unused contexts
- Removing the in-memory cache of the deal store. Badger is really fast
  and this simplifies the code tremendously