Require specific code for user devnet
Enforce a rule that requires a user to have the specific version of
filecoin deployed to user devnet when connecting to user devnet.

Note: It is still possible to connect to the user devnet without the
correct code by specifying the SHA in the FILECOIN_OVERRIDE_BUILD_SHA
environment variable during build. This solution is not intended to be
secure, only to minimize user error.

Additionally:
- Clean up some redundant or repetitive code in daemon.Init
- Bolster checks for using multiple `devnet` flags at once