localplugin: Read output in go routines (#2469)
If the data returned from stdout exceeds that of the kernal buffer,
reading from stderr will block forever because the stdout buffer will
not drain and the command will deadlock waiting on stderr to close.
Moving the reading into a go routine allows for the data to be buffered
in user-space. This is a temporary fix to enable working with large data
until #1701 can be resolved.