Proofs Mode in Genesis Block
Add the proofs mode (which also currently controls sector size) to the
genesis block via the StorageMarketActor. This makes the proofs mode a
property of the blockchain (and network) one is using instead of being
something that is based on an environment variable set at runtime.

In order to make this change, we create two versions of the genesis
block on a build of filecoin: one with the "test" proofs mode and one
with the "live" proofs mode to be used in production. The choice of the
genesis block will choose which proofs mode your node will be using.