Replace hardcoded test timeout and parallelism flags with defaults.
The specified parallism of 8 was counter-productive on machines with fewer cores, and the
hardcoded flag was impossible to override from the outside. This now uses the Go default
of GOMAXPROCS (which is #cores). The CI config sets parallelism explicitly to suit container
size.

The specified timeout was also impossible to override when a shorter value was wanted for development.
The CI config now sets the 30m value, but this is way more than we need.