Simplifications and cleanup to repo initialisation. (#2701)
Change localfilecoin IPTB plugin to nest the repo and sector dirs inside the IPTB dir.
Simplifications and cleanup to repo initialisation.

This removes the possibility of collision between files in the repo and anything else placed
in the IPTB directory (log files, pidfiles etc). Checks around whether a repo directory is suitable for initialisation have become stricter (requiring it to be empty).

This is in preparation for inserting a symlink as the repo path #2684.