Change TipSet to be an immutable list, ordered by block ticket.
Block ticket is the canonical traversal order when a tipset is processed. This change updates
the critical traversal in consensus/expected.go to match that (previously
iterated in sorted CID order). This is most of the resolution to #2310.

The representation is more memory efficient and can be iterated without allocating an
intermediate slice. An empty/undefined tipset is now harder to construct and explicitly
undefined; a few methods now always return nil error and I intend to remove the error return
value from them in a follow-up.

SortedCidSet remains as the "key" type for a tipset. In a follow-up, I intend to also
upgrade this to an immutable structure.