Split deal proposal context
The timeout context in deal proposal was being used to cover two
different cases that should be handled independently.

First we need to verify that we are connected to the miner, and secondly
we need to have our payment channel mined into a block. Everything else
can use the regular context.