Remove lookback from chain sampling for PoSt challenge seed. (#3453)
Rational PoSt has no lookback, the chain is sampled at the opening of the proving window.
The spec is unclear about what to do if this is a null tipset, this PR takes the previous
non-empty tipset. This is intended to make sense along with #3450, which waits for the
chain to settle before beginning a PoSt.

This removes PoSt challenge generation from the scope of #1872.