feat(proofscache): prepare go-filecoin for env variable change
As per https://github.com/filecoin-project/rust-fil-proofs/issues/741,
which has recently been completed and is live in rust-fil-proofs master,
the FILECOIN_PARAMETER_CACHE has been replaced with
FIL_PROOFS_PARAMETER_CACHE. That change will not be visible in
go-filecoin until the go-sectorbuilder submodule SHA is updated. To
ensure that go-filecoin is ready for that change (when the time comes to
update that submodule), we add the environment variable to Docker and
CircleCI configurations. The old environment variable can be deleted
once the cutover is complete.